qaoa changes
