+++ pennylane/qaoa/__init__.py
@@ -15,5 +15,4 @@ r"""
 Imports all of the functionality from within the QAOA submodule
 """
 
-from .mixers import *
-from .cost import *
+from .qaoa import *
