+++ pennylane/qaoa/qaoa.py
@@ -12,14 +12,59 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-This file contains functions that generate mixer Hamiltonians for use
-in QAOA workflows.
+This file contains built-in functions for defining QAOA workflows
 """
 from collections.abc import Iterable
 import networkx
 import pennylane as qml
 from pennylane.wires import Wires
-from pennylane.qaoa.utils import check_iterable_graph
+
+#############################################################
+
+
+def check_iterable_graph(graph):
+    """ Checks if a graph supplied in 'Iterable format' is valid
+
+        Args:
+            graph (Iterable): The graph that is being checked
+    """
+
+    for i in graph:
+
+        if not isinstance(i, Iterable):
+            raise ValueError(
+                "Elements of `graph` must be Iterable objects, got {}".format(type(i).__name__)
+            )
+        if len(i) != 2:
+            raise ValueError(
+                "Elements of `graph` must be Iterable objects of length 2, got length {}".format(
+                    len(i)
+                )
+            )
+        if i[0] == i[1]:
+            raise ValueError("Edges must end in distinct nodes, got {}".format(i))
+
+    if len({tuple(i) for i in graph}) != len(graph):
+        raise ValueError("Nodes cannot be connected by more than one edge")
+
+
+def get_nodes(graph):
+    """Gets the nodes of an iterable graph
+
+    Args:
+            graph (Iterable): The graph from which the nodes are obtained
+    Returns:
+            List of nodes contained in the graph
+    """
+
+    node_set = set()
+    for i in graph:
+        node_set.update([i[0], i[1]])
+
+    return list(node_set)
+
+
+############ Defines Built-In Mixer Hamiltonians ############
 
 
 def x_mixer(wires):
